+++ src/commonMain/kotlin/csense/kotlin/tests/assertions/testExtensions.kt
@@ -1,15 +1,26 @@
 package csense.kotlin.tests.assertions
 
-//contains extensions for this test module not for user code
+//contains extensions for this test module (stored away in a namespace called "helpers")
+
+public data class TestHelpers<T>(val value: T)
+
+public inline val <T> T.helpers: TestHelpers<T>
+    get() = TestHelpers(this)
+
+public fun <T> TestHelpers<T>?.simpleClassNameOrDash(): String {
+    this ?: return "-"
+    return this::class.simpleName ?: "-"
+}
+
 /**
  * If this is empty, empty is returned, otherwise the given string is
  * @receiver String
  * @param ifNotEmptyString String
  * @return String
  */
-internal fun String.useIfNotEmptyOrThis(ifNotEmptyString: String): String {
-    if (this.isEmpty()) {
-        return this
+public fun TestHelpers<String>.useIfNotEmptyOrThis(ifNotEmptyString: String): String {
+    if (value.isEmpty()) {
+        return value
     }
     return ifNotEmptyString
 }