+++ changelog.md
@@ -1,7 +1,12 @@
 # 0.0.60
 - added
     - assertDispatcher<Default, Main, Unconfined, IO>
+    - assertNotByEquals
+    - error(s) to when using assertNull / assertNotNull wrongly (eg on a non-nullable value)
 
+- changed
+  - assertIs(apply) to accept null, which of cause will fail the test
+  - 
 # 0.0.59
 
 - using kotlin 1.7.20